Working W32 build!
Changed type of lastMidiNote_ again to prevent conflict in W32
Only possible to build in Visual Studio 2008 with
  DirectX March 2008
  SDL1.2-13
These need to be linked in Solution Folders

For future reference:
Building in MinGW / Bash is still broken with the following issues
W32Midi.h is missing
GUIFactory.h has obscure errors (abstract implementation something)
WSDLGUIWindowImp doesn't build as a consequence
Linker cannot find SDL or SDL_main
Had to disble Hydra due to Dealloc not declared in this scope

Fix suggestions in https://github.com/djdiskmachine/LittleGPTracker/issues/25#issuecomment-1382974358

+++ sources/Application/Instruments/SampleInstrument.cpp
@@ -24,7 +24,7 @@ bool SampleInstrument::useDirtyDownsampling_ = false;
 
 #define SHOULD_KILL_CLICKS false
 
-int8_t SampleInstrument::lastMidiNote_[SONG_CHANNEL_COUNT]= {
+int SampleInstrument::lastMidiNote_[SONG_CHANNEL_COUNT]= {
 	-1,-1,-1,-1,-1,-1,-1,-1
 } ;
