+++ sources/Application/Instruments/SampleInstrument.h
@@ -103,7 +103,7 @@ private:
        bool dirty_ ;
 	   TableSaveState tableState_ ;
 	   
-	   static int8_t lastMidiNote_[SONG_CHANNEL_COUNT] ;
+	   static int lastMidiNote_[SONG_CHANNEL_COUNT] ;
 	   static fixed lastSample_[SONG_CHANNEL_COUNT][2] ;
 	   static fixed feedback_[SONG_CHANNEL_COUNT][FB_BUFFER_LENGTH*2] ;
